create table & form

+++ database/migrations/2024_02_19_005254_create_parameters_table.php
@@ -15,7 +15,7 @@ class CreateParametersTable extends Migration
     {
         Schema::create('parameters', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('id_parameter');
+            $table->integer('id_parameter');
             $table->string('name_parameter');
             $table->timestamps();
         });
