+++ database/migrations/2024_02_19_061616_create_componenchecks_table.php
@@ -15,7 +15,7 @@ class CreateComponenchecksTable extends Migration
     {
         Schema::create('componenchecks', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('id_componencheck');
+            $table->integer('id_componencheck');
             $table->string('name_componencheck');
             $table->timestamps();
         });
